requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS and environment management
django-cors-headers==4.3.1
python-decouple==3.8

# For deployment (optional - only install when deploying)
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0